native/voice_sidecar/requirements.txt
numpy>=1.26
sounddevice>=0.4.6
webrtcvad-wheels>=2.0.14
vosk>=0.3.45

